common/pyproject.toml
[tool.poetry]
name = "common"
version = "0.1.0"
description = "Lineage2 Interlude game server emulator"
authors = ["Yury Sokov <yury@yurzs.dev>"]
license = "MIT"
packages = [
    {include = "common", from = "."},
]

[tool.poetry.dependencies]
python = "^3.10"
aiojsonapi = "^0.3.1633554756"
motor = "^3.0.0"
typeguard = "^2.13.3"
python-dotenv = "^0.19.2"
pydantic = "^1.10.7"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0", "poetry>=1.1.12"]
build-backend = "poetry.core.masonry.api"
